+++ shared/ui/common/contact-dialog/disclaimer.tsx
@@ -5,7 +5,7 @@ const Disclaimer: React.FC = () => (
     By entering your mobile number and submitting, you consent to receive text messages from Lux at the number provided. 
     Message and data rates may apply. Message frequency varies. 
     You can unsubscribe at any time by replying STOP. 
-    View our <a href='/assets/pdf/LUX-Privacy-Policy.pdf'>Privacy Policy</a> and <a href='/assets/pdf/LUX-NFT-Terms-and-Conditions.pdf'>Terms & conditions</a>.
+    View our <a href='/assets/standard-docs/LUX-Privacy-Policy.pdf'>Privacy Policy</a> and <a href='/assets/standard-docs/LUX-NFT-Terms-and-Conditions.pdf'>Terms & conditions</a>.
   </div>
 )
