--- shared/ui/components.json
@@ -1,16 +1,0 @@
-{
-  "$schema": "https://ui.shadcn.com/schema.json",
-  "style": "default",
-  "tailwind": {
-    "config": "tailwind.config.js",
-    "css": "src/style/globals.css",
-    "baseColor": "slate",
-    "cssVariables": true 
-  },
-  "rsc": true,
-  "tsx": true,
-  "aliases": {
-    "utils": "@/util",
-    "components": "@/primitives"
-  }
-}
