+++ shared/ui/index.ts
@@ -18,3 +18,4 @@ export {
 } from './types'
 
 export * from './primitives'
+export * from './next'
