+++ shared/ui/next/index.ts
@@ -1,0 +1,11 @@
+import { determineDeviceMiddleware } from './determine-device-middleware'
+import NotFoundMDX from './not-found-content.mdx'
+import NotFound from './not-found'
+import RootLayout from './root-layout'
+
+export {
+  determineDeviceMiddleware,
+  NotFoundMDX,
+  NotFound,
+  RootLayout
+}
