+++ shared/ui/types/index.ts
@@ -12,6 +12,7 @@ import { type Breakpoint, Breakpoints } from './breakpoints'
 import type TShirtSize from './t-shirt-size' 
 import type { TShirtDimensions, Dimensions } from './dimensions' 
 import type Icon from './icon'
+import type SiteConf from './site-conf'
 
 import {
   type ContactInfo,
@@ -31,6 +32,7 @@ export {
   type TShirtDimensions,
   Breakpoints,
   type ContactInfo,
-  type ContactInfoFields
+  type ContactInfoFields,
+  type SiteConf,
 }
