+++ shared/ui/types/site-conf.ts
@@ -1,0 +1,18 @@
+import type LinkDef from './link-def'
+
+interface SiteConf {
+  title: string
+  template: string
+  desc: string
+  mainNav: {
+    short?: LinkDef[]
+    aux?: LinkDef[]
+    full: LinkDef[]
+  }
+  footer: LinkDef[][] 
+  currentAs?: string
+}
+
+export {
+  type SiteConf as default
+}
