+++ sites/market/public/assets/standard-icons/apple-touch-icon.png
