+++ sites/market/public/assets/standard-icons/favicon-32x32.png
