+++ sites/market/public/assets/standard-icons/favicon.ico
