+++ sites/market/src/app/[slug]/page.tsx
@@ -6,6 +6,8 @@ import type ProductDetailBlock from '@/blocks/def/product-detail-block'
 import ProductDetailBlockComponent from '@/blocks/components/product-detail-block'
 import { products } from '@/content'
 
+import conf from '../../conf/site'
+
 type Props = {
   params: { slug: 'silver' | 'gold' | 'coin' | 'credit' | 'validator' | 'pass' | 'uranium' }
   searchParams?: { [key: string]: string | string[] | undefined }
@@ -48,7 +50,7 @@ const ProductPage = ({ params, searchParams }: Props) => {
       <ProductDetailBlockComponent block={product} videoSize={agent === 'phone' ? 'md' : undefined}/>
     </Main>
     <div className='border-t'></div>
-    <Footer className='max-w-screen-2xl w-full pt-16 lg:mx-auto ' />
+    <Footer conf={conf} className='max-w-screen-2xl w-full pt-16 lg:mx-auto ' />
   </>)
 }
