+++ sites/market/src/app/not-found.tsx
@@ -1,16 +1,9 @@
-import React  from 'react'
+import React from 'react'
 
-import { Main, Footer } from '@luxdefi/ui/common'
-import { ApplyTypography } from '@luxdefi/ui/primitives'
-import NotFoundMDX from '@/content/not-found.mdx'
+import NotFoundCommon from '@luxdefi/ui/next/not-found'
 
-const NotFound: React.FC  = () => (<>
-  <Main className='h-[700px]'>
-    <ApplyTypography className='mt-[200px] flex flex-col md:gap-8 '>
-      <NotFoundMDX />
-    </ApplyTypography>
-  </Main>
-  <Footer />
-</>)
+import conf from '../conf/site'
+
+const NotFound: React.FC = () => (<NotFoundCommon conf={conf}/>)
 
 export default NotFound
